[CCUBE-2183][GZ] refactor source provider
