+++ tools/storybook-argtypes/services/index.ts
@@ -11,4 +11,4 @@ export { JsDocMetadataExtractor } from "./jsdoc-metadata-extractor";
 export { StoryRegistryGenerator } from "./story-registry-generator";
 export { TypeDependencyResolver } from "./type-dependency-resolver";
 export { TypeFormattingService } from "./type-formatting-service";
-export { TypeScriptSourceAnalyzer } from "./typescript-source-analyzer";
+export { TypeScriptSourceProvider } from "./typescript-source-provider";
